Implemented the synchronous API so that attaching the kernel debugger works
* Split the enable() and disable() functions in 2 so that we can bring
up only the debugger and not use the device for normal networking, plus
state transitions between the 3 enabled states.  * Implemented the
synchronous, polling sendPacket() and receivePacket() functions.  * Also
modified the interrupt filter and standard send/receive completion
handlers to interact with the debugger, but this turns out to have been
unnecessary; the interrupt handler never fires while the debugger is
active, so neither does the secondary handler. I will strip this code
back out in the next revision.  * Kernel debugging now works well in
VirtualBox (4.1) when the virtio-net adapter is the primary one in the
system. So far, I have only tested with a bridged setup, but I suspect
host-only will work too.