Removed the unnecessary interrupt handler <-> debugger interplay code
* The interrupt handler no longer cares about the debugger, as it will
never be invoked when the debugger is active.  * I've left some defense
against accidentally freeing the debugger's transmit packet in there,
just in case.  * I'm still not sure if freePacket() is safe in the
debugger code, so when dequeueing transmitted non-debugger packets, we
should really collect them until the debugger finishes and free them all
then. I've added #warnings in the relevant places.