New version 0.9.1 (Beta 2 for 1.0) with better transmit performance
* Ran some benchmarks to compare virtio to VirtualBox's Intel Gigabit
device.  * Transmit speed seems to be slightly (but statistically
significantly >95) affected by the number of physical data segments per
packets, so I've made this a tweakable parameter in the info.plist  *
fewer data segments seem to be better (presumably because the queue
doesn't fill up as fast - coalescing is obviously cheap), so I've
defaulted the value to 1  * We seem to outperform the "Intel" device
comfortably.